Don't use uint8_t for bitfields
Fixing a build failure using armcc.

+++ include/mbedtls/ssl_internal.h
@@ -319,9 +319,9 @@ struct mbedtls_ssl_handshake_params
 
         struct mbedtls_ssl_hs_buffer
         {
-            uint8_t is_valid      : 1;
-            uint8_t is_fragmented : 1;
-            uint8_t is_complete   : 1;
+            unsigned is_valid      : 1;
+            unsigned is_fragmented : 1;
+            unsigned is_complete   : 1;
             unsigned char *data;
             size_t data_len;
         } hs[MBEDTLS_SSL_MAX_BUFFERED_HS];
